V0.2.0 完成数据集构建与数据加载器的构建代码，目前能够创建数据集并加载训练样本； 解决问题：     数据加载后的images和labels分别是tuple of tensor和list of tensor, 而不是     tensor。通过自定义collate_fn函数中的torch.stack解决;
    CrossEntropyLoss中的数据类型不匹配问题, 通过label.long()解决;
todo:
    CustomDataset类还需要继续完善与精简;
    修改模型前向传播与训练过程中的bug

+++ models/losses/cross_entropy_loss.py
@@ -280,7 +280,7 @@ class CrossEntropyLoss(nn.Module):
         # Note: for BCE loss, label < 0 is invalid.
         loss_cls = self.loss_weight * self.cls_criterion(
             cls_score,
-            label,
+            label.long(),
             weight,
             class_weight=class_weight,
             reduction=reduction,
